src/requirements.txt
python>=3.9
pyext
numpy
openai
pyyaml
